refactor(query): isolate currency SQL expressions
