+++ src/services/query_service/app/repositories/transaction_repository.py
@@ -12,6 +12,7 @@ from portfolio_common.database_models import (
     Portfolio,
     Transaction,
 )
+from portfolio_common.domain.currency import normalize_currency_code
 from portfolio_common.logging_utils import operation_log_extra
 from portfolio_common.utils import async_timed
 from sqlalchemy import asc, desc, func, select
@@ -19,7 +20,7 @@ from sqlalchemy.ext.asyncio import AsyncSession
 from sqlalchemy.orm import joinedload
 
 from ..application.transaction_query import TransactionLedgerFilters, TransactionLedgerQuerySpec
-from .currency_codes import currency_code_sql_expr, normalize_currency_code
+from .currency_query_expressions import currency_code_sql_expr
 from .date_filters import start_of_day, start_of_next_day
 from .identifier_normalization import normalize_security_id
